feat: auth/authorization and notify

+++ app/Http/Livewire/Todo/Delete.php
@@ -4,9 +4,11 @@ namespace App\Http\Livewire\Todo;
 
 use App\Models\Todo;
 use Livewire\Component;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
     public Todo $todo;
 
     public function render()
@@ -16,6 +18,8 @@ class Delete extends Component
 
     public function destroy()
     {
+        $this->authorize('delete', $this->todo);
+
         $this->todo->delete();
 
         $this->emitTo(\App\Http\Livewire\Todo::class, 'todo::deleted');
